feat : implemented an API for the upcoming events
the style is the same but the events are coming from PredinctHQ API
was one of the requirements for the project.
The filtering is limited so I just set up events like expos or conferences online or in iceland
they are clickable although it links to a google search
(the free option was limited)
the other main API for events were not free, this one is for 14 days.
when the API calls doesn't work, it has a fallback on events that are hard coded for tech companies visos in Iceland